compile post view earlier
