+++ crates/api/api_crud/src/post/read.rs
@@ -13,11 +13,11 @@ use lemmy_db_schema::{
 };
 use lemmy_db_views_community::CommunityView;
 use lemmy_db_views_local_user::LocalUserView;
-use lemmy_db_views_post::{
+use lemmy_db_views_post::PostView;
+use lemmy_db_views_search_combined::{
   api::{GetPost, GetPostResponse},
-  PostView,
+  impls::SearchCombinedQuery,
 };
-use lemmy_db_views_search_combined::impls::SearchCombinedQuery;
 use lemmy_db_views_site::SiteView;
 use lemmy_utils::error::{LemmyErrorType, LemmyResult};
