+++ crates/db_views/post/Cargo.toml
@@ -23,14 +23,12 @@ full = [
   "i-love-jesus",
   "lemmy_db_schema/full",
   "lemmy_db_schema_file/full",
-  "lemmy_db_views_community/full",
   "lemmy_db_views_vote/full",
 ]
 ts-rs = [
   "dep:ts-rs",
   "lemmy_db_schema/ts-rs",
   "lemmy_db_schema_file/ts-rs",
-  "lemmy_db_views_community/ts-rs",
   "lemmy_db_views_vote/ts-rs",
 ]
 
@@ -38,7 +36,6 @@ ts-rs = [
 lemmy_db_schema = { workspace = true }
 lemmy_utils = { workspace = true, optional = true }
 lemmy_db_schema_file = { workspace = true }
-lemmy_db_views_community = { workspace = true }
 lemmy_db_views_vote = { workspace = true }
 diesel = { workspace = true, optional = true }
 diesel-async = { workspace = true, optional = true }
